move pages to pages directory
