--- src/components/TransactionPage.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function TransactionPage() {
-    return <h1>This is transaction page</h1>
-}
