+++ src/components/pages/TransactionPage.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Transaction from '../Transaction';
+
+export default function TransactionPage() {
+    return (
+        <>
+            <h1>This is transaction page</h1>
+            <Transaction/>
+        </>
+    )
+}
